+++ frontend/wailsjs/go/models.ts
@@ -27,6 +27,20 @@ export namespace plugins {
 
 export namespace profiles {
 	
+	export class Subscription {
+	    filter: string;
+	    qos: number;
+	
+	    static createFrom(source: any = {}) {
+	        return new Subscription(source);
+	    }
+	
+	    constructor(source: any = {}) {
+	        if ('string' === typeof source) source = JSON.parse(source);
+	        this.filter = source["filter"];
+	        this.qos = source["qos"];
+	    }
+	}
 	export class ConnectionProfile {
 	    id: string;
 	    name: string;
@@ -39,8 +53,9 @@ export namespace profiles {
 	    username: string;
 	    password: string;
 	    keepAlive: number;
-	    subFilter: string;
-	    subQos: number;
+	    subscriptions: Subscription[];
+	    subFilter?: string;
+	    subQos?: number;
 	
 	    static createFrom(source: any = {}) {
 	        return new ConnectionProfile(source);
@@ -59,9 +74,28 @@ export namespace profiles {
 	        this.username = source["username"];
 	        this.password = source["password"];
 	        this.keepAlive = source["keepAlive"];
+	        this.subscriptions = this.convertValues(source["subscriptions"], Subscription);
 	        this.subFilter = source["subFilter"];
 	        this.subQos = source["subQos"];
 	    }
+	
+		convertValues(a: any, classs: any, asMap: boolean = false): any {
+		    if (!a) {
+		        return a;
+		    }
+		    if (a.slice && a.map) {
+		        return (a as any[]).map(elem => this.convertValues(elem, classs));
+		    } else if ("object" === typeof a) {
+		        if (asMap) {
+		            for (const key of Object.keys(a)) {
+		                a[key] = new classs(a[key]);
+		            }
+		            return a;
+		        }
+		        return new classs(a);
+		    }
+		    return a;
+		}
 	}
 
 }